refactor: readSubmissions
