+++ packages/synthetic-chain/src/lib/core-eval-support.ts
@@ -1,4 +1,5 @@
-// @ts-check
+import * as fsp from 'node:fs/promises';
+
 import {
   Far,
   makeMarshal,
@@ -141,4 +142,18 @@ export const ensureISTForInstall = async (
   );
   log({ wantMinted });
   await mintIST(addr, sendValue, wantMinted, giveCollateral);
+};
+
+export const readSubmissions = async () => {
+  const files = await fsp.readdir('submission');
+  const names = files.filter(f => f.endsWith('.js')).map(f => f.slice(0, -3));
+  const buildAssets = {} as Record<string, BundleInfo>;
+  for (const name of names) {
+    const evals = [{ permit: `${name}-permit.json`, script: `${name}.js` }];
+    const content = await fsp.readFile(`submission/${name}.js`, 'utf8');
+    const bundleIds = content.matchAll(/b1-[a-z0-9]+/g);
+    const bundles = Array.from(bundleIds).map(id => `${id}.json`);
+    buildAssets[name] = { evals, bundles };
+  }
+  return buildAssets;
 };
