+++ proposals/b:zoe1/performActions.ts
@@ -15,30 +15,15 @@ import { voteLatestProposalAndWait } from '@agoric/synthetic-chain/src/lib/commo
 import { dbTool } from '@agoric/synthetic-chain/src/lib/vat-status.js';
 import { type WebCache } from '@agoric/synthetic-chain/src/lib/webAsset.js';
 import {
-  type BundleInfo,
   bundleDetail,
   ensureISTForInstall,
   flags,
   getContractInfo,
   loadedBundleIds,
   txAbbr,
+  readSubmissions,
 } from '@agoric/synthetic-chain/src/lib/core-eval-support.js';
 import { step } from '@agoric/synthetic-chain/src/lib/logging.js';
-
-// TODO move into core-eval-support
-const readSubmissions = async () => {
-  const files = await fsp.readdir('submission');
-  const names = files.filter(f => f.endsWith('.js')).map(f => f.slice(0, -3));
-  const buildAssets = {} as Record<string, BundleInfo>;
-  for (const name of names) {
-    const evals = [{ permit: `${name}-permit.json`, script: `${name}.js` }];
-    const content = await fsp.readFile(`submission/${name}.js`, 'utf8');
-    const bundleIds = content.matchAll(/b1-[a-z0-9]+/g);
-    const bundles = Array.from(bundleIds).map(id => `${id}.json`);
-    buildAssets[name] = { evals, bundles };
-  }
-  return buildAssets;
-};
 
 /**
  * URLs of assets, including bundle hashes (to be) agreed by BLD stakers
